+++ tests/tests_pytorch/deprecated_api/test_remove_1-10.py
@@ -15,6 +15,7 @@
 import pytest
 
 from pytorch_lightning import Trainer
+from pytorch_lightning.core.mixins.device_dtype_mixin import DeviceDtypeModuleMixin
 from pytorch_lightning.demos.boring_classes import BoringModel
 from pytorch_lightning.overrides import LightningDistributedModule, LightningParallelModule
 from pytorch_lightning.overrides.base import unwrap_lightning_module
@@ -69,3 +70,11 @@ def test_v1_10_deprecated_unwrap_lightning_module_sharded():
 def test_v1_10_deprecated_on_colab_kaggle_func():
     with pytest.deprecated_call(match="The function `on_colab_kaggle` has been deprecated in v1.8.0"):
         on_colab_kaggle()
+
+
+def test_v1_10_deprecated_device_dtype_module_mixin():
+    class MyModule(DeviceDtypeModuleMixin):
+        pass
+
+    with pytest.deprecated_call(match="mixins.DeviceDtypeModuleMixin` has been deprecated in v1.8.0"):
+        MyModule()